basic threading implemented to ODMR
Basic threading example has been implemented to the ODMR window. Currently when ODMR window opens, a thread starts to count up until it reaches a maximum number or the ODMR window closes.  When the function completes, it emits a finished signal to a slot that prints the result, will also print out error to terminal if it doesnt not run