Threading w/ callback/error/finished signal emitters and slots
Threading added in that can record data and plot after the fact or plot in real time have been added. The real time plotting works using a callback function to sample the data in the thread and plot it, without interrupting the data taking.

ODMR sweeps can be started and then stopped at any time and any data taken up to that point will be plotted and can be analysed.