removed error in freq. plotting
Moved the trigger command for the RF source to start the sweep to be the last thing that is sent, to make sure that all other variables are assigned and that the LIA is primed for data acquisition. This removes some time descrepency. Further, found that you need a ratio of 1:0.25:10 for the dwell time, burst duration and number of samples per burst to keep a good consistent measurement. Too many samples can cause a longer read time relative to the dwell time in the freq sweep and also having a too short burst duration relative to the dwell time can cause overlap of measurements.