finished logic checks for equipment connections
Using scalar scan just checks that the printer/RF/LIA are connected.
Using feedback or vector scans checks this as well as checks if there is at least 1 row of frequencies and gradients in the table.
Vector scans can now use how many freqs. given in the table instead of being fixed to just 4.