data & image exporting added
When scans finish, data can be exported to a h5 format as well as .PNG images of all the data arrays to user defined folder. Images are scaled to 640 x 640 size using cubic interpolation just for viewing purposes. Raw data in the .h5 file should be used for real analysis.